Documentei a classe Aposta

+++ src/modelos/Aposta.java
@@ -16,8 +16,8 @@ public class Aposta {
 	 * Constrói uma aposta a partir dos parâmetros apostados e previsao
 	 * que são Strings e o parâmetro quantia que é um inteiro.
 	 * @param apostador String que é o nome de quem fez a aposta.
-	 * @param quantia Inteiro 
-	 * @param previsao
+	 * @param quantia Inteiro que é o quanto em centavos foi apostado.
+	 * @param previsao String que é previsão do cenário.
 	 */
 	public Aposta(String apostador, int quantia, String previsao) {
 		this.apostador = apostador;
@@ -25,18 +25,34 @@ public class Aposta {
 		this.previsao = previsao;
 	}
 
-	public double getQuantia() {
+	/**
+	 * Método retorna o valor do atributo quantia.
+	 * @return Inteiro que é a quantia que foi apostada.
+	 */
+	public int getQuantia() {
 		return quantia;
 	}
 
+	/**
+	 * Método retorna o valo do  atributo apostador.
+	 * @return String que é o nome de quem apostou.
+	 */
 	public String getApostador() {
 		return apostador;
 	}
 
+	/**
+	 * Método retorna o valor do atributo previsao.
+	 * @return String que é a previsão sobre o cenário.
+	 */
 	public String getPrevisao() {
 		return previsao;
 	}
 	
+	/**
+	 * Método retorna uma representação textual de uma aposta.
+	 * @return String que é auma representação textual da aposta.
+	 */
 	public String toString() {
 		return (this.apostador + " - R$" + (this.quantia/100.0) + " - " + this.previsao);
 	}
